requirements.txt
easydict==1.9
numpy==2.2.4
tqdm==4.31.1 
torchvision==0.21.0
torch==2.6.0
opencv_python==4.11.0
Pillow==11.1.0
tensorboardX==2.0 
face-recognition==1.3.0
cvzone==1.6.1
python-dotenv==1.1.0
supabase-py==2.15.0

